feat-#16: move types to domain
