+++ src/item/application/commands/impl/notify-item-owner.command.ts
@@ -1,4 +1,4 @@
-import { ItemEventType } from '../../../types';
+import { ItemEventType } from '../../../domain/types';
 
 export class NotifyItemOwnerCommand {
   constructor(public readonly event: ItemEventType) {}
