+++ src/item/domain/events/impl/item-owner-notified.event.ts
@@ -1,6 +1,6 @@
 import { IEvent } from '@nestjs/cqrs';
 
-import { ItemEventType } from '../../../types';
+import { ItemEventType } from '../../types';
 
 export class ItemOwnerNotifiedEvent implements IEvent {
   constructor(public readonly event: ItemEventType) {}
