+++ src/item/domain/repositories/item.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 
 import { CreateItemDto, UpdateItemDto } from '../../dtos';
 import { Item } from '../models';
-import { ItemEventType } from '../../types';
+import { ItemEventType } from '../types';
 
 @Injectable()
 export class ItemRepository {
